feat: update endpoints and create new published-projects service
