+++ app/hooks/published-projects/types.ts
@@ -1,0 +1,6 @@
+// usePublishedProjects
+export interface UsePublishedProjectsProps {
+  search?: string;
+  sort?: string,
+  filters?: Record<string, unknown>
+}
