+++ app/pages/community/projects/[pid]/index.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router';
 import { withUser } from 'hoc/auth';
 import { withPublishedProject } from 'hoc/projects';
 
-import { usePublishedProject } from 'hooks/projects';
+import { usePublishedProject } from 'hooks/published-projects';
 
 import ProjectDetail from 'layout/community/published-projects/detail';
 import Footer from 'layout/footer';
